[feat] duplicate nested and nested_accelerate function from vector.rs to nested.rs
